Rename evaluator.py to vector.py
