+++ nand/test_codegen.py
@@ -512,4 +512,4 @@ def cycles_per_second():
 def test_speed():
     cps = cycles_per_second()
     print(f"Measured speed: {cps:0,.1f} cycles/s")
-    assert cps > 100000
+    assert cps > 100_000
