+++ nand/vector.py
